adjust after cherry-pick 71c6ca5

+++ imagine/tools/random_seed.py
@@ -50,4 +50,4 @@ def ensemble_seed_generator(size):
     """
     log.debug('@ random_seed::ensemble_seed_generator')
     # the uint32 is defined by the random generator's capasity
-    return np.random.randint(low=1, high=np.uint32(-1)//3, size=np.uint(size))
+    return np.random.randint(low=1, high=np.uint32(-1)//3, size=np.uint64(size))
